# openapi.yaml
openapi: 3.0.0
info:
  title: Dream API
  description: Api
  contact:
    name: Backend Developer
  version: 2.0.0.1
servers:
- url: https://virtserver.swaggerhub.com/sad-41a/dream/2.0.0.1
  description: SwaggerHub API Auto Mocking
- url: http://localhost:8080/api/v1
  description: Local Server
paths:
  /auth/register:
    post:
      tags:
      - Auth
      summary: Регистрация донора (Личный кабинет)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auth_register_body"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
  /auth/login:
    post:
      tags:
      - Auth
      summary: Вход (Админ или Донор)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auth_login_body"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
  /auth/me:
    get:
      tags:
      - Auth
      summary: Получить профиль текущего пользователя
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      security:
      - bearerAuth: []
  /dreams:
    get:
      tags:
      - Dreams (Public)
      summary: Список всех мечтаний (с фильтрами)
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - active
          - completed
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 9
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200"
    post:
      tags:
      - Dreams (Admin)
      summary: Создать новую карточку мечты
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/dreams_body"
      responses:
        "201":
          description: Created
      security:
      - bearerAuth: []
  /dreams/{id}:
    get:
      tags:
      - Dreams (Public)
      summary: Детальная информация о мечте
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dream"
    put:
      tags:
      - Dreams (Admin)
      summary: Редактировать мечту
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Dream"
      responses:
        "200":
          description: Updated
      security:
      - bearerAuth: []
    delete:
      tags:
      - Dreams (Admin)
      summary: Удалить мечту
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Deleted
      security:
      - bearerAuth: []
  /dreams/{id}/donors:
    get:
      tags:
      - Dreams (Public)
      summary: "Список тех, кто помог этой мечте"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DonationPublic"
  /donations/pay:
    post:
      tags:
      - Payment
      summary: Создать платеж
      description: "Если user_id не передан, считается анонимным. Возвращает ссылку на эквайринг."
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/donations_pay_body"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_1"
  /donations/my:
    get:
      tags:
      - User Cabinet
      summary: История моих пожертвований
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/inline_response_200_2"
      security:
      - bearerAuth: []
  /news:
    get:
      tags:
      - News
      summary: Получить список новостей
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/News"
    post:
      tags:
      - News (Admin)
      summary: Опубликовать новость
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/News"
      responses:
        "201":
          description: Created
      security:
      - bearerAuth: []
  /reports:
    get:
      tags:
      - Reports
      summary: Финансовые отчеты
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      tags:
      - Reports (Admin)
      summary: Загрузить отчет (PDF)
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/reports_body"
      responses:
        "201":
          description: Uploaded
      security:
      - bearerAuth: []
  /upload:
    post:
      tags:
      - Upload
      summary: Загрузка картинок (для админки)
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/upload_body"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_3"
      security:
      - bearerAuth: []
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: integer
    AuthToken:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        avatar_url:
          type: string
        role:
          type: string
          enum:
          - user
          - admin
        total_donated:
          type: number
          description: Сумма всех пожертвований пользователя
    Dream:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: "Мечта Вани: увидеть море"
        slug:
          type: string
          description: Для красивой ссылки в React
          example: vanya-more
        short_description:
          type: string
        full_description:
          type: string
          description: HTML контент
        target_amount:
          type: number
        collected_amount:
          type: number
        status:
          type: string
          enum:
          - active
          - completed
          - frozen
        cover_image:
          type: string
        gallery_images:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
      description: Карточка мечты (сбора)
    DonationPublic:
      type: object
      properties:
        amount:
          type: number
        donor_name:
          type: string
          example: Алексей
        created_at:
          type: string
          format: date-time
        is_anonymous:
          type: boolean
      description: Публичная информация о донате (для списка 'Помогли')
    News:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        preview_text:
          type: string
        content:
          type: string
        image_url:
          type: string
        published_at:
          type: string
          format: date-time
    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        year:
          type: integer
          example: 2024
        month:
          type: integer
          example: 5
        file_url:
          type: string
          description: Ссылка на PDF
        title:
          type: string
    auth_register_body:
      required:
      - email
      - full_name
      - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        full_name:
          type: string
    auth_login_body:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    inline_response_200:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Dream"
        total:
          type: integer
    dreams_body:
      required:
      - target_amount
      - title
      type: object
      properties:
        title:
          type: string
        short_description:
          type: string
        full_description:
          type: string
        target_amount:
          type: number
        cover_image:
          type: string
    donations_pay_body:
      required:
      - amount
      - dream_id
      type: object
      properties:
        amount:
          type: number
        dream_id:
          type: string
          format: uuid
        email:
          type: string
        is_anonymous:
          type: boolean
        comment:
          type: string
    inline_response_200_1:
      type: object
      properties:
        payment_url:
          type: string
    inline_response_200_2:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        dream_title:
          type: string
        date:
          type: string
          format: date-time
        status:
          type: string
    reports_body:
      type: object
      properties:
        file:
          type: string
          format: binary
        title:
          type: string
        year:
          type: integer
    upload_body:
      type: object
      properties:
        file:
          type: string
          format: binary
    inline_response_200_3:
      type: object
      properties:
        url:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT